feat: add NEAR review storage
