+++ agents/review-agent.ts
@@ -1,6 +1,6 @@
 import { Review } from '../types';
 import { assertTonChain } from '../services/chain';
-import { addReview, getReviews } from '@/features/reviews/services/tonReviews';
+import { addReview, getReviews } from '@/features/reviews/services/nearReviews';
 import ordersAgent from './orders-agent';
 import productsAgent from './products-agent';
 import storesAgent from './stores-agent';
