+++ src/features/reviews/services/tonReports.ts
@@ -1,4 +1,4 @@
-import { setValue, listValues, removeValue } from '@/services/tonKvStore';
+import { setValue, listValues, removeValue } from '@/services/nearKvStore';
 import { Report } from '@/types';
 import { requireEnv } from '@/utils/appConfig';
 import { assertTonChain } from '@/services/chain';
